move resolver helper methods to separate module
